==> sepsis_project/DataFormattingScript/2b_fix_dates.ipynb <==
#this file formats the raw NUR vital files while also fixing the dates in them

import pandas as pd
import numpy as np
import math

df1 = pd.read_excel(
     '../original_data/ectas_edm_vitals.xlsx',
     engine='openpyxl', sheet_name="ectas_vitals"
)

df1

del df1['td.locat']
del df1["Temp_Source"]
del df1["Interv"]

rows = len(df1.index)
df1

df2 = pd.DataFrame(data=[], columns= ['pt_identifier', 'Date', 'Time', 'BP1', 'BP2', 'Pulse', 'Temp', 'RR', 'O2Sat'])

df2

rows

def extract_vitals(df1, i):
    p_id = df1.iloc[i][0]
    date = df1.iloc[i][1]
    time = df1.iloc[i][2]
    pulse = df1.iloc[i][4]
    rr = df1.iloc[i][5]
    o2sat = df1.iloc[i][6]
    temp = df1.iloc[i][7]
    
    # print (df1.iloc[i][3])
    # print (type(df1.iloc[i][3]))
    if isinstance(df1.iloc[i][3], str):
        pre_bp = df1.iloc[i][3]
        if type(pre_bp) == float and math.isnan(pre_bp)==False:
            manual_bp = str(pre_bp)
            x = manual_bp.split("/")
            if len(x)==2:
                bp1 = float(x[0])
                bp2 = float(x[1])
            else:
                bp1 = None
                bp2 = None
        elif type(pre_bp) == str:
            manual_bp = str(pre_bp)
            if "E" in manual_bp:
                manual_bp = manual_bp.replace("E", "")
                manual_bp = manual_bp.strip()
            x = manual_bp.split("/")
            if len(x)==2:
                bp1 = float(x[0])
                bp2 = float(x[1])
            else:
                bp1 = None
                bp2 = None
        else:
            bp1 = None
            bp2 = None

    elif df1.iloc[i][8] and df1.iloc[i][9] and math.isnan(df1.iloc[i][8])==False and math.isnan(df1.iloc[i][9])==False:
        bp1 = df1.iloc[i][8]
        bp2 = df1.iloc[i][9]

    else:
        bp1 = None
        bp2 = None
    

    # print (df1.iloc[i][8])
    # print (df1.iloc[i][9])
    # print (math.isnan(df1.iloc[i][8]))
    # print (math.isnan(df1.iloc[i][9]))
    
    p_vitals = [p_id, date, time, bp1, bp2, pulse, temp, rr, o2sat]
    return p_vitals

i = 0
pt_identifier = []
date = []
time = []
bp1 = []
bp2 = []
pulse = []
temp = []
rr = []
o2sat = []
while i < rows:
    p_vitals = extract_vitals(df1, i)
    # df2 = df2.append({'pt_identifier': p_vitals[0], 'Date': p_vitals[1],
    #                   'Time': p_vitals[2], 'BP1': p_vitals[3], 'BP2': p_vitals[4],
    #                   'Pulse': p_vitals[5], 'Temp': p_vitals[6], 'RR': p_vitals[7],
    #                   'O2Sat': p_vitals[8]}, ignore_index=True)
    #print(p_vitals)
    pt_identifier.append(p_vitals[0])
    date.append(p_vitals[1])
    time.append(p_vitals[2])
    bp1.append(p_vitals[3])
    bp2.append(p_vitals[4])
    pulse.append(p_vitals[5])
    temp.append(p_vitals[6])
    rr.append(p_vitals[7])
    o2sat.append(p_vitals[8])
    i += 1
  
    


mydict = {"pt_identifier":pt_identifier, "Date": date, "Time": time, "BP1": bp1, "BP2": bp2, "Pulse": pulse, "Temp":temp, "RR": rr, "O2Sat":o2sat}
 #'Time': p_vitals[2], 'BP1': p_vitals[3], 'BP2': p_vitals[4],
    #                   'Pulse': p_vitals[5], 'Temp': p_vitals[6], 'RR': p_vitals[7],
    #                   'O2Sat': p_vitals[8]}, ignore_index=True)

df3 = pd.DataFrame(mydict)

df3

date = df3.iloc[1][1]
print(date)
date = str(date)
new_date = date[0:6] + "20" + date[6:8]
new_date


i = 0
new_date = []
while i<rows:
    date = df3.iloc[i][1]
    date = str(date)
    date = date[0:6] + "20" + date[6:8]
    new_date.append(date)
    i+=1

new_date

df3["Date"] = new_date

df3

df3.to_csv('../FormattedData/2_redone_date.csv')#formatted_good_nur_vitalsall.csv


# run time approx 5 minutes for 75,000 rows
