==> sepsis_project/DataFormattingScript/7a_sort_by_time_normalize expanded.ipynb <==
#this file normalizes the 'date' into relative minutes
#for the expanded version

import pandas as pd

df1 = pd.read_csv(
     '../FormattedData/31_fixed.csv',
)

df1

del df1["Unnamed: 0"]

df1

import datetime

print(df1.iloc[312757][1])
print(type(df1.iloc[312757][1]))
print((df1.iloc[1][1]))
print(type(df1.iloc[1][1]))


rows = len(df1)
rows

df1['Date'] = df1['Date'].astype(str)


def create_new_pt_identify():
    c = 1
    new_idx = 0
    temp_date = []
    temp_date.append(0)
    string = True
    while c<rows:
        date = df1.iloc[c][1]
        # print(date)
        if ("-" in date or "/" in date) and string == False:
            # print(1)
            string = True
            new_idx+=1
            temp_date.append(new_idx)
        elif "-" not in date and "/" not in date and string == True:
            # print(2)
            string = False
            new_idx+=1
            temp_date.append(new_idx)
        elif df1.iloc[c][0] != df1.iloc[c-1][0]:
            new_idx+=1
            temp_date.append(new_idx)
        else:
            # print(4)
            temp_date.append(new_idx)
        c+=1
    print (len(temp_date))
    df1["temp_date"] = temp_date
    df1.head(50)

df1

test = "0012"
float(test)

#create new identifier every time it switches
#then just handle it as separate events; do it normally with this handler here; consecutive will be right
def weird_date_handler(df1, i, consecutive):
    current_row = i
    weird_date = []
    weird_time = []
    weird_normalized = []

    for x in range(consecutive):
        weird_date.append(df1.iloc[current_row][1])
        weird_time.append(df1.iloc[current_row][2])

    minutes = []
    
    #change times into minutes
    for y in range(len(weird_time)):
        a = float(weird_time[y])
        b = int(a)

        if b<100:
            minutes.append(b)

        elif b<1000 and b>=100:
            c = b%100
            d = int(str(b)[0])
            minutes.append( (d*60) + c)
        
        else:
            e = b%100
            f = int(str(b)[0:2])
            minutes.append( (f*60) + e)

    z=0

    #add 24 hours = 1440 minutes when the date changes
    while z<(len(weird_date)-1):
        if weird_date[z]!=weird_date[z+1]:
            minutes[z+1][0] = minutes[z+1][0]+1440
        z+=1

    #find the min value and subtract it from everything to make everything relative to the 0 min
    min = minutes[0]
    for xy in range(1, len(minutes)):
        if minutes[xy]<min:
            min = minutes[xy]

    for yz in range(len(minutes)):
        minutes[yz] = minutes[yz]-min
        
    return minutes

test = "Nan"
print(type(test))
a = float(test)
print(a)

import numpy as np
rows = len(df1)
i = 0
while i<rows:
    # if np.isnan(df1.iloc[i][1]):
    #     print(i)
    #     print("date")
    if np.isnan(df1.iloc[i][2]):
        print(i)
        print("time") 

    i+=1

#remove those rows

df1.reset_index(inplace = True, drop = True)

df1[432730:432740]

#get date and time into lists per patient;
#do NOT change order
#find lowest date in that list
#make another list with same indexing of relative time (lowest is 0)(units in minutes --> 0, 10, 20, 70, 100)
#combine all these lists (append to main)
#put it in df unsorted
#sort df
def date_to_number(df1, i, consecutive):
    current_row = i
    date_list = []
    time_list = []

    for x in range(consecutive):
        date_list.append(df1.iloc[current_row][1])
        time_list.append(df1.iloc[current_row][2])
        current_row+=1
    
    date_num_list = []
    date_type_list = []

    for y in range(len(date_list)):
        #print(date_list[y])
        date = str(date_list[y])
        date = date.split(" ")[0]
        if "-" in date:
            x = date.split("-")
            date_int = [int(x[2]), int(x[1]), int(x[0])]

        elif "/" in date:
            x = date.split("/")
            date_int = [int(x[0]), int(x[1]), int(x[2])]

        #need in format day, month, year
        date_num_list.append(date_int)

    time_num_list = []
    for z in range(len(time_list)):
        #print(time_list[z])
        a = float(time_list[z])
        b = int(a)

        if b<100:
            time_num_list.append([0,b])

        elif b<=1000 and b>100:
            c = b%100
            d = int(str(b)[0])
            time_num_list.append([d,c])
        
        else:
            e = b%100
            f = int(str(b)[0:2])
            time_num_list.append([f,e])
    
    # print(len(date_num_list))
    # print(len(time_num_list))
    date_time_list = []
    for xy in range(consecutive):
        #print(date_num_list[xy])
        new_date = datetime.datetime(date_num_list[xy][2], date_num_list[xy][1], date_num_list[xy][0], time_num_list[xy][0], time_num_list[xy][1])
        date_time_list.append(new_date)
        
    #print(datetime.datetime(2020, 5, 17))
    min = date_time_list[0]
    low_index = 0

    for yz in range(1, len(date_time_list)):
        if date_time_list[yz]<min:
            min = date_time_list[yz]
            low_index = yz
    
    normalized_time = []

    for xyz in range(len(date_time_list)):
        date1 = date_time_list[xyz]
        date0 = date_time_list[low_index]
        result_days = int( (date1-date0).days)
        result_seconds = int( (date1-date0).seconds)
        good_result = float( (((86400*result_days)+result_seconds)/60) )
        normalized_time.append(good_result)

    return normalized_time



def count_consecutive(df1, i, last_idx):
    count = 1

    while i<rows-1 and df1.iloc[i][0] == df1.iloc[i+1][0] :
        count += 1
        i += 1
        # print(i)
        if i == last_idx:
            break
    
    return count    

final_array = []
i=0
while i<rows:
    consecutive = count_consecutive(df1, i, rows-1)
    #print(i)
    each_array = date_to_number(df1, i, consecutive)
    for t in each_array:
        final_array.append(t)
    i+=consecutive
    


print(len(final_array))

len(final_array)
df1["Date"] = final_array
df1

#df1.sort_values(by=["new_date"], inplace = True)

del df1["Time"]

df1

df1.to_csv('../FormattedData/32.csv')#vitals_info_lactate_WBC_sorted_pt_and_time, normalized too
