==> sepsis_project/DataFormattingScript/8_sort_by_patient_then_time2.ipynb <==
# this file sorts again by patient and then by time

import pandas as pd



df1 = pd.read_csv(
     '../FormattedData/32.csv',
)

df1 = df1.rename(columns={'Date': 'new_date'})


df1.head(50)

del df1["Unnamed: 0"]

df2 = df1.sort_values(['pt_identifier', 'new_date'], ascending=[True, True])


df2.head(50)

df2.to_csv('../FormattedData/33.csv')#vitals_info_lactate_WBC_normalized_time_sorted_pt_and_time_stage3


df2
