==> sepsis_project/DataFormattingScript/9_drop_unneeded_columns_duplicate_patient_info_into_lab_rows.ipynb <==
#this file drops unneeded columns, and puts the patient info into lab rows

import pandas as pd



df1 = pd.read_csv(
     '../9.csv',
)

# 'Unnamed: 0', 'Unnamed: 0.1', 'Age_Code',
df1.drop(['Unnamed: 0', 'Unnamed: 0.1', 'Age_Code','Diag_Code'], axis = 1, inplace = True)

df1.head(50)

import math
ageDict = {}
genderDict = {}
weightDict = {}
diagnosisDict = {}

print(diagnosisDict.get('H1012043'))

df1.at[1, 'Age_Number'] = 81

rows = df1.shape[0]

# make dict for age // works
i = 0
while i < rows:
    if (df1.iloc[i][0] not in ageDict) and (math.isnan(df1.iloc[i][8]) == False):
        ageDict[df1.iloc[i][0]] = df1.iloc[i][8]
    i = i + 1

# make dict for gender // works
i = 0
while i < rows:
    if (df1.iloc[i][0] not in genderDict) and (isinstance(df1.iloc[i][9], str) == True):
        genderDict[df1.iloc[i][0]] = df1.iloc[i][9]
    i = i + 1

# make dict for weight // works
i = 0
while i < rows:
    if (df1.iloc[i][0] not in weightDict) and (math.isnan(df1.iloc[i][10]) == False):
        weightDict[df1.iloc[i][0]] = df1.iloc[i][10]
    i = i + 1

# make dict for diagnosis // works
i = 0
while i < rows:
    if (df1.iloc[i][0] not in diagnosisDict):
        if isinstance(df1.iloc[i][11], str) == True:
            diagnosisDict[df1.iloc[i][0]] = df1.iloc[i][11]
    i = i + 1

# fill in blanks
i = 0
while i < rows:
    # age
    if (math.isnan(df1.iloc[i][8]) == True) and (df1.iloc[i][0] in ageDict):
        df1.at[i, 'Age_Number'] = ageDict.get(df1.iloc[i][0])
    
    # gender
    if (isinstance(df1.iloc[i][9], str) == False) and (df1.iloc[i][0] in genderDict):
        df1.at[i, 'Gender'] = genderDict.get(df1.iloc[i][0])
    
    # weight
    if (math.isnan(df1.iloc[i][10]) == True) and (df1.iloc[i][0] in weightDict):
        df1.at[i, 'Weight'] = weightDict.get(df1.iloc[i][0])
    
    # diagnosis
    if (isinstance(df1.iloc[i][11], str) == False):
        if df1.iloc[i][0] in diagnosisDict:
            df1.at[i, 'Diag_desc'] = diagnosisDict.get(df1.iloc[i][0])
        else:
            df1.at[i, 'Diag_desc'] = 'Non sepsis patient'

    
    i = i + 1

df1.head(80)

df1.to_csv('../FormattedData/10.csv')#vitals_info_lactate_WBC_normalized_time_sorted_pt_and_time_duplicate_patient_info_into_lab_rows_stage4
