==> sepsis_project/DataFormattingScript/12_separate_old_data.ipynb <==
#this file looks at the time and separates them if the value is over 10,000 minutes away (1 week)

import pandas as pd

df1 = pd.read_csv(
     '../FormattedData/37.csv'
)

df1

del df1["Unnamed: 0"]

df1

rows = len(df1)
rows

print (df1.iloc[0][1])


i=1
date_list = []
date_list.append(df1.iloc[0][0])
days = 3
time_threshold = days*1440
while i<rows-2:
    value1 = df1.iloc[i][0]
    value0 = df1.iloc[i-1][0] 

    if (value1-value0) > time_threshold:
        subtracted = df1.iloc[i][0]
        date_list.append(0.0)
        new_idx = i+1

        while new_idx<rows and df1.iloc[new_idx][0] != 0.0 and (df1.iloc[new_idx][0]-df1.iloc[new_idx-1][0])<time_threshold:
            changed_v = df1.iloc[new_idx][0]
            date_list.append(changed_v - subtracted)
            new_idx += 1
            i += 1
    else:
        date_list.append(value1)
    

    i+=1
    


date_list[0:50]

del df1["new_date"]
df1["new_date"] = date_list

df1 = df1[['new_date', 'BP1', 'BP2', 'Pulse', 'Temp', 'RR', 'O2Sat',
       'Age_Number', 'Male', 'Female']]

df1


df1.head(50)

df1.to_csv('38.csv')
