==> sepsis_project/DataFormattingScript/14_take_out_single_entries.ipynb <==
#this file deletes single rows from the dataset; ie. if there is no time series; ie. if there is only 1-2 entries of that instance
#ie. time: new patient: 0, 5, 10, 15, 20, 25; new patient: 0, 1; new patient, 0, 10, 20, 30, 40, 50; take out the 0,1 guy

import pandas as pd
#import modin.pandas as pd

df1 = pd.read_csv(
     '../FormattedData/38.csv'
)

df1

del df1["Unnamed: 0"]

df1

rows = len(df1)
rows

print (df1.iloc[65][3])
print (type(df1.iloc[65][3]))
#row column


i=0
to_drop = []


while i<rows-2:
    value1 = df1.iloc[i][0]
    value2 = df1.iloc[i+1][0]
    value3 = df1.iloc[i+2][0]

    if value1 == 0.0 or value1 == 0:
        if value2 == 0.0 or value2 == 0:
            to_drop.append(i)

        elif value3 == 0.0 or value3 == 0:
            to_drop.append(i)
            to_drop.append(i+1)

    i+=1
    
print(len(to_drop))
# print(to_drop)

df1 = df1.drop(to_drop)
df1

df1.head(50)

df1 = df1.reset_index(drop=True)
df1.head(50)

len(df1)

df1.to_csv('../FormattedData/39.csv')
