==> sepsis_project/DataFormattingScript/16_prep_for_ml.ipynb <==
#this file deletes the postadmit categories

import pandas as pd
#import modin.pandas as pd

df1 = pd.read_csv(
     '../FormattedData/16.csv'
)

df1

del df1["Unnamed: 0"]

df1.head(10)

rows = len(df1)
rows

del df1["Most_responsible"]
del df1["pre_admit"]
del df1["post_admit"]
del df1["secondary"]

df1

print (df1.iloc[65][1])
print ((df1.iloc[65][3]))
#row column


df1

df2 = df1.fillna(-1)

df2.to_csv('../FormattedData/17a.csv') 


df2
