==> sepsis_project/DataFormattingScript/17_a_average_time_calc.ipynb <==
#this file calculates average minutes in each time step

import pandas as pd
#import modin.pandas as pd

df1 = pd.read_csv(
     '../FormattedData/39.csv'
)

df1

del df1["Unnamed: 0"]

df1.head(10)

rows = len(df1)
rows

df1.iloc[2][1] #row, column

del df1["pt_identifier"]

df1["new_date"] = pd.to_numeric(df1["new_date"])


i = 1
times = []
while i<rows:
    time = df1.iloc[i][0]
    prev_time = df1.iloc[i-1][0]
    if time != 0:
        times.append(time-prev_time)
    i+=1


average = sum(times)/len(times)

df1.to_csv("../FormattedData/34.csv")

average
#is the average time step
