==> sepsis_project/DataFormattingScript/17_c_determine_missing_values.ipynb <==
#this file calaculatas how many and the % of missing values in the dataset

import pandas as pd
#import modin.pandas as pd

df1 = pd.read_csv(
     '../FormattedData/39.csv'
)

df1

df1.head(30)

df1.columns = ['index', 'new_date', 'BP1', 'BP2', 'Pulse', 'Temp', 'RR', 'O2Sat', 'Age_Number', 'Gender']

df1

new_row = pd.DataFrame({'index':0, 'new_date':0.00, 'BP1':126.00,
                        'BP2':93.00, 'Pulse':115.00,
                        'Temp':36.20, 'RR':18.00, 'O2Sat':98.00, 'Age_Number': 46, 'Gender': "M"},
                                                            index =[0])


new_row

df1 = pd.concat([new_row, df1]).reset_index(drop = True)
df1.head(5)

df1.head(20)

del df1["index"]

df1.head(30)

rows = len(df1)
rows

df1.drop(df1.tail(1).index,inplace=True) # drop last n rows

df1

del df1["Unnamed: 0"]
df1

df1.iloc[2][1] #row, column

new_date = []
bp1 = []
bp2 = []
hr = []
temp = []
rr = []
o2sat = []
age = []
# lactate = []
# wbc = []

df2 = df1.fillna(-1)
df2

df2.to_csv("../FormattedData/35.csv")

rows = len(df1)
rows

i = 1


while i<rows:
    if df1.iloc[i][0] == -1:
        new_date.append(0)
    else:
        new_date.append(1)

    if df1.iloc[i][1] == -1:
        bp1.append(0)
    else:
        bp1.append(1)

    if df1.iloc[i][2] == -1:
        bp2.append(0)
    else:
        bp2.append(1)

    if df1.iloc[i][3] == -1:
        hr.append(0)
    else:
        hr.append(1)

    if df1.iloc[i][4] == -1:
        temp.append(0)
    else:
        temp.append(1)

    if df1.iloc[i][5] == -1:
        rr.append(0)
    else:
        rr.append(1)

    if df1.iloc[i][6] == -1:
        o2sat.append(0)
    else:
        o2sat.append(1)

    if df1.iloc[i][7] == -1:
        age.append(0)
    else:
        age.append(1)

    # if df1.iloc[i][10] == -1:
    #     lactate.append(0)
    # else:
    #     lactate.append(1)

    # if df1.iloc[i][11] == -1:
    #     wbc.append(0)
    # else:
    #     wbc.append(1)

    i+=1


def compute_missing(alist):
    size = len(alist)
    count = 0
    for i in alist:
        if i == 0:
            count+=1
    print(float(count)/float(size))

df1.to_csv("../FormattedData/36.csv")

compute_missing(new_date)
compute_missing(bp1)
compute_missing(bp2)
compute_missing(hr)
compute_missing(temp)
compute_missing(rr)
compute_missing(o2sat)
compute_missing(age)
# compute_missing(lactate)
# compute_missing(wbc)
#is the average number of steps
