==> sepsis_project/DataFormattingScript/17_secondprep_for_ml.ipynb <==
#this file deletes the first 3 categories and flags no lab values as -1

import pandas as pd
#import modin.pandas as pd

df1 = pd.read_csv(
     '../FormattedData/17a_deleted_outliers_manually.csv'
)

df1

del df1["Unnamed: 0"]

df1.head(10)

rows = len(df1)
rows

del df1["pt_identifier"]
del df1["positive_diag"]
del df1["negative_diag"]


df1

df1.loc[df1['Lactate_adjusted'] == 0.0, 'Lactate_adjusted'] = -1


df1

df1.loc[df1['WBC_adjusted'] == 0.0, 'WBC_adjusted'] = -1


df1

df1.to_csv('../FormattedData/pre_ml_2.csv')
