==> sepsis_project/DataFormattingScript/18_prep_for_baseline_1.ipynb <==
#this file deletes the first 3 categories and flags no lab values as -1

import pandas as pd
#import modin.pandas as pd

df1 = pd.read_csv(
     '../FormattedData/pre_ml_3.csv'
)

df1

del df1["Unnamed: 0"]

df1.head(10)

rows = len(df1)
rows

del df1["Lactate_adjusted"]
del df1["WBC_adjusted"]


new_date = []
bp1 = []
bp2 = []
hr = []
temp = []
rr = []
o2sat = []
age = []
lactate = []
wbc = []

df1

averaging_time = 120

def average_and_append(df1, idx):
    #find next 0 to find entries in between
    temp_count = idx+1
    sample_size = 1
    while df1.iloc[temp_count][0] != 0:
        temp_count+=1
        sample_size+=1

    for i in range(sample_size):
        
        pass
    pass

i = 0 
while i<rows:
    if df1.iloc[i][0] == 0:
        #if newdate == 0
        average_and_append(df1, i)
        
    i+=1

df1

df1.to_csv('../FormattedData/pre_ml_2.csv')
